EW-356 removing old file builder
